Migrate NftModal, NftGridModal, TransactionModal
