+++ apps/mobile-wallet/src/features/nftsDisplay/NftModal.tsx
@@ -26,62 +26,70 @@ import AppText from '~/components/AppText'
 import Button from '~/components/buttons/Button'
 import { ModalScreenTitle, ScreenSection } from '~/components/layout/Screen'
 import NFTImage, { NFTImageProps } from '~/components/NFTImage'
-import { ModalContent, ModalContentProps } from '~/features/modals/ModalContent'
+import BottomModal from '~/features/modals/BottomModal'
+import { ModalContent } from '~/features/modals/ModalContent'
+import withModalWrapper from '~/features/modals/withModalWrapper'
 import { useAppSelector } from '~/hooks/redux'
 import { BORDER_RADIUS_SMALL, DEFAULT_MARGIN } from '~/style/globalStyle'
 
-type NFTModalProps = ModalContentProps & Pick<NFTImageProps, 'nftId'>
+type NftModalProps = Pick<NFTImageProps, 'nftId'>
 
 const attributeGap = 12
 const windowWidth = Dimensions.get('window').width
 const nftFullSize = windowWidth - DEFAULT_MARGIN * 4
 const attributeWidth = (nftFullSize - attributeGap) / 2
 
-const NFTModal = ({ nftId, ...props }: NFTModalProps) => {
+const NftModal = withModalWrapper<NftModalProps>(({ id, nftId }) => {
   const nft = useAppSelector((s) => selectNFTById(s, nftId))
   const { t } = useTranslation()
 
   if (!nft) return null
 
   return (
-    <ModalContent verticalGap {...props}>
-      <ScreenSection>
-        <ModalScreenTitle>{nft.name}</ModalScreenTitle>
-      </ScreenSection>
-      <ScreenSection>
-        <NFTFullSizeImage nftId={nftId} size={nftFullSize} />
+    <BottomModal
+      id={id}
+      noPadding
+      Content={(props) => (
+        <ModalContent {...props} verticalGap>
+          <ScreenSection>
+            <ModalScreenTitle>{nft.name}</ModalScreenTitle>
+          </ScreenSection>
+          <ScreenSection>
+            <NFTFullSizeImage nftId={nftId} size={nftFullSize} />
 
-        {nft.description && (
-          <NFTDescriptionContainer>
-            <AppText color="secondary" size={16}>
-              {nft.description}
-            </AppText>
-          </NFTDescriptionContainer>
-        )}
-        {nft.attributes && nft.attributes.length > 0 && (
-          <AttributesGrid>
-            {nft.attributes.map((attribute) => (
-              <Attribute key={attribute.trait_type} style={{ width: attributeWidth }}>
-                <AttributeType color="tertiary" semiBold>
-                  {attribute.trait_type}
-                </AttributeType>
-                <AttributeValue semiBold>{attribute.value}</AttributeValue>
-              </Attribute>
-            ))}
-          </AttributesGrid>
-        )}
-      </ScreenSection>
+            {nft.description && (
+              <NFTDescriptionContainer>
+                <AppText color="secondary" size={16}>
+                  {nft.description}
+                </AppText>
+              </NFTDescriptionContainer>
+            )}
+            {nft.attributes && nft.attributes.length > 0 && (
+              <AttributesGrid>
+                {nft.attributes.map((attribute) => (
+                  <Attribute key={attribute.trait_type} style={{ width: attributeWidth }}>
+                    <AttributeType color="tertiary" semiBold>
+                      {attribute.trait_type}
+                    </AttributeType>
+                    <AttributeValue semiBold>{attribute.value}</AttributeValue>
+                  </Attribute>
+                ))}
+              </AttributesGrid>
+            )}
+          </ScreenSection>
 
-      {!nft.image.startsWith('data:image/') && (
-        <ScreenSection>
-          <Button title={t('View full size')} onPress={() => openBrowserAsync(nft.image)} />
-        </ScreenSection>
+          {!nft.image.startsWith('data:image/') && (
+            <ScreenSection>
+              <Button title={t('View full size')} onPress={() => openBrowserAsync(nft.image)} />
+            </ScreenSection>
+          )}
+        </ModalContent>
       )}
-    </ModalContent>
+    />
   )
-}
+})
 
-export default NFTModal
+export default NftModal
 
 const NFTFullSizeImage = styled(NFTImage)`
   flex: 1;
